requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==3.0.5
charset-normalizer==3.3.2
cloudinary==1.40.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.4
cryptography==42.0.5
daphne==3.0.2
Django==5.0.4
django-cors-headers==4.3.1
django-docs==0.3.3
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==1.7.1
pyOpenSSL==24.1.0
pytest==8.1.1
python-dotenv==1.0.1
requests==2.31.0
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
zope.interface==6.3
